+++ src/nl/hannahsten/texifyidea/inspections/latex/typesetting/LatexMissingGlossaryReferenceInspection.kt
@@ -6,6 +6,7 @@ import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
+import nl.hannahsten.texifyidea.index.LatexProjectStructure
 import nl.hannahsten.texifyidea.index.NewSpecialCommandsIndex
 import nl.hannahsten.texifyidea.inspections.InsightGroup
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionBase
@@ -21,7 +22,7 @@ import nl.hannahsten.texifyidea.util.toTextRange
 class LatexMissingGlossaryReferenceInspection : TexifyInspectionBase() {
     override val inspectionGroup = InsightGroup.LATEX
     override val inspectionId = "MissingGlossaryReference"
-    override fun getDisplayName() = "Missing glossary reference"
+    override fun getDisplayName() = "Missing glossary or acronym reference"
 
     private val nameLetterRegex = "[^a-zA-Z]+".toRegex()
 
@@ -41,9 +42,12 @@ class LatexMissingGlossaryReferenceInspection : TexifyInspectionBase() {
         val descriptors = descriptorList()
         val regexes = extractedNames.map { nameLetters ->
             val nameLetterRegex = nameLetters.toRegex()
-            val glsRegex = "\\\\gls[^{]+\\{($nameLetters)}".toRegex()
+            // Both glossaries and acronym packages provide acronymsh
+            val glsRegex = "\\\\(?:gls|ac)[^{]+\\{($nameLetters)}".toRegex()
             nameLetterRegex to glsRegex
         }
+        val libraries = LatexProjectStructure.getFilesetDataFor(file)?.libraries ?: emptySet()
+
         // Unfortunately the lowest level we have is a block of text, so we have to do a text-based search
         file.traverseTyped<LatexNormalText>().forEach { textElement ->
             val text = textElement.text
@@ -51,14 +55,20 @@ class LatexMissingGlossaryReferenceInspection : TexifyInspectionBase() {
                 val correctOccurrences = glsRegex.findAll(text).mapNotNull { it.groups.firstOrNull()?.range }
                 val allOccurrences = nameLettersRegex.findAll(text).map { it.range }
                 allOccurrences.filter { !correctOccurrences.contains(it) }.forEach { range ->
+                    // The command is different for each package, but the idea is the same
+                    val fixes = listOf(Pair("glossaries.sty", "\\gls"), Pair("acronym.sty", "\\ac")).filter { it.first in libraries }
+                        .map { AddGlsFix(it.second) }
+                        .toTypedArray()
+                        .ifEmpty { arrayOf(AddGlsFix("\\gls")) }
+
                     descriptors.add(
                         manager.createProblemDescriptor(
                             textElement,
                             range.toTextRange(),
-                            "Missing glossary reference",
+                            "Missing glossary or acronym reference",
                             ProblemHighlightType.GENERIC_ERROR_OR_WARNING,
                             isOntheFly,
-                            AddGlsFix(),
+                            *fixes,
                         )
                     )
                 }
@@ -67,13 +77,13 @@ class LatexMissingGlossaryReferenceInspection : TexifyInspectionBase() {
         return descriptors
     }
 
-    private class AddGlsFix : LocalQuickFix {
-        override fun getFamilyName() = "Add \\gls command"
+    private class AddGlsFix(private val command: String) : LocalQuickFix {
+        override fun getFamilyName() = "Add $command command"
 
         override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
             val range = descriptor.textRangeInElement
             val newText = descriptor.psiElement.text.replaceRange(range.endOffset, range.endOffset, "}")
-                .replaceRange(range.startOffset, range.startOffset, "\\gls{")
+                .replaceRange(range.startOffset, range.startOffset, "$command{")
 
             val newElement = LatexPsiHelper(project).createFromText(newText).firstChild
             descriptor.psiElement.replace(newElement)